Added control to wrap WebView2 logic
